requirements.txt
PyMuPDF
EbookLib
Pillow
beautifulsoup4 